### 2024-08-28 #7156 @editorialbot
[PRE REVIEW]: QuantizedSystemSolver: A discontinuous ODE system 2 solver for Julia.
**Submitting author:** @mongibellili (<a href="http://orcid.org/0000-0003-1230-5488">elmongi elbellili</a>)
**Repository:** https://github.com/mongibellili/QuantizedSystemSolver.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
