### 2024-10-17 #7156 @mongibellili
Sure, I searched for all with "julia" and "numerical analysis", and removed those working on many other specific fields. I found these reviewers:
-Edenhofer
-rafaelbailo
-DanielVandH
-lucaferranti
-lamBOOO
-ManuelBehrendt
-vlc1
-victorapm
-dannys4
-RahulSundar
-mvaldes
-cmsmlab
-EdwardBerman
-simone-silvestri
-ChrisRackauckas
-josemanuel22
-ajinkyakadu
-sandreza
-ziyixi
-hemanthgrylls
-joeybernard
-mseri
