### 2024-10-23 7937f1a
update the docs: internals explanation
