### 2024-10-23 6fba5e5
Update dependencies.md
