+++ docs/src/dependencies.md
@@ -20,9 +20,7 @@ such as addsub, subsub, and mulT are designed to operate directly on existing da
 mathematical functions: Similar to arithmetic operations, new mathematical functions are
 designed to leverage in-place operations, which reduces the overhead associated with creating
 new memory for intermediate results. For instance, functions like exp, log, sin, and cos, sqrt, power iterate over the input Taylor series and apply the corresponding operation without allocating additional arrays.
-By implementing these new methods, the code significantly reduces the number of allocations, improving both execution speed and resource management, while still maintaining the
-flexibility and functionality required for complex mathematical operations.
-transformation of expressions: The transformF function is designed to translate userdefined mathematical expressions into optimized forms that leverage the custom arithmetic and
-function implementations. By traversing the expression tree with the prewalk function, it identifies operations such as addition, subtraction, multiplication, division, and specific mathematical functions (like exponential and logarithmic functions). For each identified operation, the code modifies the expression to call specialized versions (e.g., subT, addT, mulT) that avoid memory allocation, enhancing performance during calculations. Additionally, it tracks the number of caches needed for these operations, which are essential for efficient computation of Taylor series.
+By implementing these new methods, the code significantly reduces the number of allocations, improving both execution speed and resource management, while still maintaining the flexibility and functionality required for complex mathematical operations.
+transformation of expressions: The transformF function is designed to translate userdefined mathematical expressions into optimized forms that leverage the custom arithmetic and function implementations. By traversing the expression tree with the prewalk function, it identifies operations such as addition, subtraction, multiplication, division, and specific mathematical functions (like exponential and logarithmic functions). For each identified operation, the code modifies the expression to call specialized versions (e.g., subT, addT, mulT) that avoid memory allocation, enhancing performance during calculations. Additionally, it tracks the number of caches needed for these operations, which are essential for efficient computation of Taylor series.
 Each time a transformation occurs, a reference to a cache is added, ensuring that a necessary
 cache is available for the optimized functions. This systematic approach effectively restructures the original expressions, maintaining their functionality while improving computational efficiency by reducing unnecessary allocations.
