### 2024-11-05 #7156 @lamBOOO
> Hi @lamBOOO do you have time to review this paper?

Yes, I can do that.
