### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#11 @lamBOOO
JOSS: Software > General
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Remove Empty Files**: Files such as `.github/workflow` are empty and should be removed.
- **Clean Up Redundant Directories**: The `build` folder appears to be a partial copy of `src`. If it is unnecessary, it should be removed.
- **Improve Documentation**: Not all functions are documented with docstrings. Comprehensive documentation would enhance usability.
- **Eliminate Unnecessary Comments**: The code contains many comments that may not be needed, for example, in `tests/results_systems.jl`. Removing them can improve code readability.
- **Simplify Build Scripts**: The `docs/make.jl` file contains unnecessary code, such as `run(`cp -r "C:/Users/belli/.julia/dev/QuantizedSystemSolver/docs/build/*"`)`. This should be cleaned up.
- **Remove Unnecessary Files**:
  - `lcov.info`
  - `src/dense/NL_integrators/workspace.code-workspace`
  - Please review the repository thoroughly to remove files that are not relevant to users.
- **Standardize Directory Naming**: Some folders within `src` use uppercase letters while others use lowercase. Consistent naming conventions should be applied.
