### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#12 @lamBOOO
JOSS: Software > Design
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Use of `quote ... end` in Problem Definition**: The `NLodeProblem` utilizes the `quote ... end` construct, resulting in an `Expr` object. For example:
  ```julia
  odeprob = NLodeProblem(quote
      name=(sysb53,)
      u = [-1.0, -2.0]
      du[1] = -20.0*u[1]-80.0*u[2]+1600.0
      du[2] = 1.24*u[1]-0.01*u[2]+0.2
  end)
  ```
  This contrasts with a more conventional approach:
  ```julia
  ode_prob = NLodeProblem(
      name = sysb53,
      u0 = [-1.0, -2.0],  # Initial condition for u
      f = function (du, u)
          du[1] = -20.0 * u[1] - 80.0 * u[2] + 1600.0
          du[2] = 1.24 * u[1] - 0.01 * u[2] + 0.2
      end
  )
  ```
  - **Concerns**: This unconventional use of `quote ... end` may not be the most elegant or efficient method in terms of type inference, robustness, and API accessibility.
  - **Recommendation**: Please explain the rationale behind this design choice or consider adopting the more standard method of using function arguments in Julia. The current approach makes it unclear what the function does with the input provided within the `quote ... end` block.
