### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#13 @lamBOOO
JOSS: Software > README
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Include Logo**: The documentation features a logo; consider adding it to the README for consistency and branding purposes.
- **Provide Usage Instructions**: The README should include instructions on how to execute the software's functionality, including the necessary `using ...` commands to import the package.
- **Fix Code Errors**: The following code snippet from the README returns an error:
  ```julia
  # Returns a vector of the values of all variables at time 0.0005 (feature not available in v1.0.1)
  sol(0.0005)
  ```
  **Error Message**:
  ```
  ERROR: MethodError: no method matching (::QuantizedSystemSolver.LightSol{2, 2})(::Float64)
  Closest candidates are:
    (::Sol)(::Int64, ::Float64)
     @ QuantizedSystemSolver ~/.julia/packages/QuantizedSystemSolver/XwCgz/src/Common/Solution.jl:93
  Stacktrace:
   [1] top-level scope
     @ REPL[17]:1
  ```
  This suggests that the feature may not be available or properly implemented. Please resolve this issue or update the README to reflect the current capabilities.
- **Mathematical Explanation of the ODE Problem**: The ODE problem presented in the README should be accompanied by a mathematical explanation. This will help users understand how to formulate their own problems and apply the software effectively.
- **Clarify Feature Availability**: The README mentions several features marked as "available only in v1.0.1." This can be confusing for users. Please update the README to reflect the current version and available features accurately.
- **Plot Mismatch**: The provided plot does not match the expected output (as per the attached image). Ensure that the plots in the README accurately represent the software's output to avoid confusion.
- **Error in Error Analysis Code**: The error analysis code provided at the end does not run successfully. Please verify and correct the code examples to ensure they work as intended.
- **Improve Image Alignment**: In the section titled "Example: Buck Circuit," the image alignment could be improved. Centering the image on a separate line would enhance the visual presentation.

![Screenshot 2024-11-07 at 16 13 55](https://github.com/user-attachments/assets/e1f1cdae-3ab6-4bde-89e4-14c476fba64a)
