### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#14 @lamBOOO
JOSS: Software > Documentation
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Enhance Index Page**: The [index page](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/) should include a clear statement of need, an overview of the package, and links to tutorials and API documentation. As the first point of contact, it should effectively introduce the software and guide users to relevant resources.
- **Detailed Mathematical Descriptions**: Where possible, include detailed mathematical descriptions of the problems being solved, particularly in tutorials. This will aid users in understanding the underlying concepts and applying the software to their own problems.
- **Improve Navigation**: The API reference and list of algorithms are only linked within the tutorial. Consider adding them to the navigation bar on the left for easier access.
  - The same applies to pages linked in the [Developer Guide](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/developerGuide/).
- **Enhance High-Level Documentation**: The high-level documentation in the [Application Programming Interface](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/interface/#Application-Programming-Interface) section needs improvement.
  - Clearly explain how to use the `NLodeProblem` API.
  - Use proper formatting for function calls, similar to examples in [DFTK](https://docs.dftk.org/stable/api/) or [Gridap](https://gridap.github.io/Gridap.jl/stable/Io/).
- **Purpose of Dependencies Page**: The [Dependencies](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/dependencies/#Dependencies) documentation page provides detailed descriptions of dependencies like *SymEngine*. While acknowledging dependencies is important, such detailed explanations may not be necessary. Consider summarizing or removing this section unless it serves a specific purpose.
- **Consolidate Repeated Links**: The "Algorithm Extension" link is repeated in the [Developer Guide](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/developerGuide/). Please consolidate or remove redundant links to streamline navigation.
- **Improve Extension Guides**: The `Problem` and `Solution` extension pages do not provide practical guidance on how to extend the software to new problems. Including simple, step-by-step examples would greatly benefit users looking to customize the software.
- **Clarify `Taylor0` Documentation**: The [Taylor0](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/Taylor0) documentation needs improvement. The purpose and functionality of this component are not clearly explained. Providing more context and examples would be helpful.
- **Organize Examples**: The [Examples](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/examples/) section could be organized into separate subpages to reduce complexity and improve user navigation.
- **Automate Example Compilation**: Consider using [Literate.jl](https://github.com/fredrikekre/Literate.jl) to automatically compile examples. This ensures that code examples execute correctly and allows users to run them directly, which is a common practice in the Julia ecosystem.
- **Exclude `build` Folder from Git**: The `build` folder within `docs` is included in the Git repository. Typically, this folder is generated automatically during continuous integration (CI) and should not be part of the repository. Excluding it can prevent unnecessary clutter and potential conflicts.
