### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#15 @lamBOOO
JOSS: Software > API Documentation
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Add Doc strings**: Please add doc strings for all functions, such as those in `src/Utils/rootfinders/SimUtils.jl`. Comprehensive documentation enhances usability and helps users understand how to utilize the API effectively.
