### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#16 @lamBOOO
JOSS: Software > Functionality
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Evaluation of Correctness**: While the software appears capable of solving ODE problems, it is challenging to assess the correctness of the implementation for more complex cases, such as stiff ODEs with events. Including a simple yet comprehensive example would greatly assist in validating the correctness and robustness of the solver. Are there any simple cases, where a comparison to exact solution might be possible?
- **Example of Exponential Decay**: I tried solving the simple ODE $\dot{x}(t) = - x(t)$ with initial condition $x(0) = 1$ (whose analytical solution is $x(t) = e^{-t}$) which works correctly:
  ```julia
  using LinearAlgebra
  using QuantizedSystemSolver

  odeprob = NLodeProblem(quote
    name = (exp_decay)
    u = [1.0]
    du[1] = -u[1]
  end)

  sol = solve(odeprob, nmliqss2(), tspan, abstol=1e-10, reltol=1e-10)
  error_norm = norm(map(x -> exp(-x), sol.savedTimes[1]) - sol.savedVars[1])
  # Output: 5.0644554862553414e-9
  ```
  Is it possible to include such a simple example that showcases the full capabilities of the software, i.e., with events, multiple solution variables which, at the same time, can be easily evaluated for correctness (like the exponential decay from above)?
