### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#17 @lamBOOO
JOSS: Software > Tests
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Reference Results in Tests**: It would be beneficial if the test suite (e.g., `test/Tyson.jl`) included comparisons of computed results against reference solutions. This would help in detecting unintended changes in the software's output due to modifications, allowing developers to investigate and address issues promptly.
  - Notably, there are commented-out sections in tests like `test/adr.jl` that reference XLSX data. It is unclear why these comparisons are commented out. Improving the code quality of the tests by reactivating and ensuring these comparisons work correctly would enhance the reliability of the testing framework.
- **Consistent Naming Conventions**: The files within the `test` folder currently have inconsistent naming conventions regarding the use of underscores and capitalization. Adopting a consistent naming scheme would improve organization and make the test suite easier to navigate and maintain.
