### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#18 @lamBOOO
JOSS: Software > Community Guidelines
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

- **Contribution Guidelines**: The project currently lacks clear guidelines for third parties who wish to contribute to the software, report issues, or seek support. Please provide detailed instructions.
  - These guidelines should be prominently featured in the README or a dedicated `CONTRIBUTING.md` file and referenced in the documentation. This will encourage community engagement and make it easier for users and developers to collaborate effectively.
