### 2024-11-08 mongibellili/QuantizedSystemSolver.jl#19 @lamBOOO
JOSS: Paper
Concerning the [JOSS review thread](https://github.com/openjournals/joss-reviews/issues/7434).

### Preamble

- **Author Affiliation Correction**: The second author's affiliation should be corrected from "Kuleuven" to "KU Leuven" to accurately reflect the institution's official name.

### Paragraph 1

- **Clarity on Engineering Systems and Stiffness**: The sentence "The growing intricacy of contemporary engineering systems, typically reduced to stiff differential equations with events, poses a difficulty in digitally simulating them using traditional numerical integration techniques" may not be clear to non-specialists.
  - **Recommendation**: Provide a brief explanation of why contemporary engineering systems are often modeled as stiff ODEs with events. For instance:
    - **Stiffness Explanation**: Explain that stiffness arises in systems where there are rapidly changing and slowly changing components, making numerical simulation challenging due to the need for very small time steps to maintain stability.
    - **Engineering Systems Context**: Clarify what is meant by "contemporary engineering systems," perhaps by providing examples such as electrical circuits, mechanical systems with shocks, or chemical reactions with rapid kinetics.
- **Precision in Describing the Quantized Approach**: The statement "It is an approach that builds the solution by updating the system variables independently as opposed to classic integration methods that update all the system variables every step" could be made more precise.
  - **Recommendation**:
    - **Mathematical Explanation**: Introduce mathematical formulations to explain the quantized state method (QSM). For example, describe how variables are updated only when they surpass certain thresholds (quantization levels), reducing the computational effort.
    - **Advantages Over Traditional Methods**: Highlight why the quantized approach is better suited for stiff ODEs compared to traditional numerical integration schemes, perhaps by discussing computational efficiency or improved stability.
- **Formatting of Citations**: The last citation in this paragraph has an extraneous period. Ensure that all citations are correctly formatted according to the chosen citation style.

### Paragraph 2

- **Word Repetition Correction**: The phrase "Julia language Julia programming language" contains a repetition.
- **Definition of an Event**: The term "event" is used without definition.
  - **Recommendation**: Provide a clear definition of an event in the context of ODEs.
- **Consistency in Capitalization**: The term "Ordinary Differential Equations" should be consistently capitalized.
- **Improvement in Sentence Structure**: The sentence "the QuantizedSystemSolver.jl is a solver that aims to efficiently solve" can be rephrased for clarity.
  - **Recommendation**: Maybe, replace it with "The `QuantizedSystemSolver.jl` package aims to efficiently solve...".

### Paragraph 3

- **Consistent Use of Capitalization**: The terms "Problem," "Algorithm," and "Solution" are capitalized without apparent reason. Think about using a monospaced font to indicate function or classes.
- **Figure Reference**: The figure should be referenced within the text using a hyperlink or a figure number.
- **Formatting of Code Elements**: The use of `NLodeProblem` should be formatted consistently.
- **Clarification on Naming Conventions**: "NLodeProblem", I find the variable name strange, why is ODE written in lowercase?
  - **Recommendation**: Consider renaming it to `NLODEProblem` or `NonlinearODEProblem` for clarity and adherence to common naming conventions.
- **Improvement of Code Examples**: The sample code could be more informative and aligned with standard Julia practices.
  - **Recommendation**:
    - **Use Real Julia Code**: Provide actual Julia code that can be executed, with comments explaining each part.
    - **Match with Mathematical Examples**: Ensure that the code corresponds to a mathematical example provided earlier in the text.
    - **Syntax Highlighting**: Use syntax highlighting for code snippets to improve readability (like in https://doi.org/10.21105/joss.04157).
    **Example**:
    ```julia
    # This is just a demonstration what I mean
    odeprob = NLodeProblem(
      quote,  # quote
      "Simulation 1",  # name
      (a=1, b=x->x^2),  # parameters
      [x, t],  # discrete and continous variables
      helper->beep(),  # helper expressions
      [(x,t)->a*sin(x)*exp(-t)],  # differential equations
      [
        (t-1, execute_event())  # do something at t==1
      ],  # if-statments for events
    )
    ```
- **Explanation of `quote ... end` Usage**: The use of `quote ... end` in the code is unconventional and may confuse readers.
  - **Recommendation**: Explain why this construct is used or consider adopting the standard method of defining functions and passing them as arguments, which is more familiar to Julia users. See also the other comment about the software design
- **Consistency in Whitespace and Formatting**:
  - Add a whitespace after the assignment operator in `sol=`.
  - Ensure consistent use of whitespace after commas, e.g., `sol(0.0005,idxs=2)` vs. `tspan = (0.0, 0.001)`.
- **Formatting of Citations**: Ensure that there is proper whitespace before citations, such as `SymEngine.jl (Team, 2015)`.
- **Reference Author Listings**: For the three references in this paragraph, clarify how the authors are listed in the reference section to maintain consistency and proper attribution. Are the authors taken from the `LICENSE` file?
- **Clarity and Precision in Conclusion**:
  - The sentence "In conclusion, the package offers extensive functionality to facilitate practical research tasks" is vague.
    - **Recommendation**: Be specific about the functionalities offered by the package, such as efficiently solving stiff ODEs with events using the quantized state method.
  - **Typographical Errors**: Correct the typo in "In Fact, Anyone" by removing the unnecessary capitalization.
  - **Simplify Complex Sentences**: The sentence "In conclusion, the package offers extensive functionality to facilitate practical research tasks, all while being lightweight and well documented to be easily used by researchers and students to efficiently model various dynamical systems with discontinuities, as well as further study and improve the newly developed QSS methods" is lengthy and complex. Please consider rewriting it.
  - **Grammar Corrections**:
    - Change "Anyone that has a different problem type" to "Anyone who has a different problem type."

### Figures

- **Descriptive Caption**: The figure should have a descriptive caption that explains its content and relevance to the text.
- **Formatting Within the Figure**:
  - In the "User Interface" node of Figure 1, ensure consistent use of whitespace and alignment for readability.

### References

- **Missing DOI**: The reference "Pietro et al." is missing a DOI.

### General Remarks

- **Statement of Need**:
  - **Recommendation**: Include a clear statement of need that illustrates the purpose of the software. This could involve presenting a challenging stiff ODE with events and demonstrating how the software effectively solves it.
- **Comparison with Other Packages**:
  - **Recommendation**:
    - Include a comparison with traditional numerical integration methods to highlight the advantages of the quantized approach (maybe in a plot?).
    - Reference other software packages in this domain, even if they are not in Julia.
      - Maybe this one (https://github.com/CIFASIS/qss-solver) if it's relevant?
- **Ongoing Research and Publications**:
  - **Recommendation**: Mention any ongoing research projects that utilize the software or recent scholarly publications enabled by it. This demonstrates the software's applicability and impact in the research community.
- **Software Archiving for Reproducibility**:
  - **Recommendation**: Archive the software using a platform like Zenodo and include a citation with a DOI in the paper. This ensures long-term accessibility and facilitates reproducibility.
