### 2024-11-09 mongibellili/QuantizedSystemSolver.jl#20 @joshuaeh
JOSS Feedback
Re: [JOSS Review](https://github.com/openjournals/joss-reviews/issues/7434)

A few notes while reviewing the JOSS guidelines:

- Would be helpful to have a link to the documentation on the GH repo page (to the right and within the readme) (https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/)
- Installation functional, but could be described better for the Julia novices
- State of the field could be clearer: contrasting what is currently done, especially within the paper: maybe an example of a stiff system and compare approaches
- RE: the v.1.01 notes: use version control so that these are not shown until it is released then just show it and note those functions in the release changelog

- **Documentation**:
  - Landing page with description (statement of need) and installation would be good
  - The examples do a good job of showing how the qwuantized approach adds vaue with the number of dots: leverage this and maybe explain the differences between approaches
  - Tutorial: would be good to have a general simple example / just show the syntax and setup of the problem before solving the first one
  - Developer Guide:
    - Implementation page is a huge plock of text and daunting
    - Implementation: Distinguish between syntax and placeholders in the example
  - Implementation/Dependencies hide the valuable API reference information
