### 2024-11-09 #7434 @joshuaeh
This is good work: handling stiff systems efficiently is useful for all kinds of downstream tasks and hopefully researchers can use and extend these methods rather than implementing their own. @lamBOOO has done a great job, I have a few points as well mostly on the documentation and presentation side noted in a new issue:

[https://github.com/mongibellili/QuantizedSystemSolver.jl/issues/20](https://github.com/mongibellili/QuantizedSystemSolver.jl/issues/20)
