+++ paper/paper.bib
@@ -24,7 +24,7 @@
   year = {2015},
 }
 @misc{SymEngine,
-  author       = {Development Team},
+  author       = {Isuru Fernando},
   title        = {SymEngine.jl},
   url = {https://github.com/symengine/SymEngine.jl},
   year = {2015},
@@ -34,4 +34,53 @@
   title        = {Improving linearly implicit quantized state system methods},
   journal = {Simulation: Transactions of the Society for Modeling and Simulation International},
   year = {2019},
+  doi     = {10.1177/0037549718766},
+}
+
+@article{qssC,
+author = {Fernández, Joaquín and Kofman, Ernesto},
+year = {2014},
+month = {06},
+pages = {782-799},
+title = {A stand-alone quantized state system solver for continuous system simulation},
+volume = {90},
+journal = {SIMULATION: Transactions of the Society for Modeling and Simulation International},
+doi = {10.1177/0037549714536255}
+}
+
+@article{Rackauckas2017,
+ author = {Rackauckas, C. and Nie, Q.},
+ title = {DifferentialEquations.jl A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+  volume = {5},
+   pages = {15-24},
+ year = {2017},
+journal = {The Journal of Open Research Software},
+}
+ 
+@Inproceedings{elbellili, 
+  author =        {Elbellili, E. and Lauwens, B. and Huybrechs, D.},
+  title =        {Investigation of Different Conditions to Detect Cycles in Linearly Implicit Quantized State Systems},
+  booktitle =    {Proceedings of the 3rd. International Conference on Computational Modeling, Simulation and Optimization},
+  year =         {2024},
+  publisher =    {IEEE Computer Society},
+}
+
+@book{fe2006a,
+  author = {Cellier, F. and Kofman , E.},
+  title = {Continuous system},
+  publisher = {Springer},
+  year = {2006},
+  language = {en},
+  address = {simulation.New York},
+}
+
+@article{buck,
+author = {Cristri, A.W.},
+year = {2017},
+month = {12},
+pages = {398-403},
+title = {Analysis and Design of Dynamic Buck Converter with Change in Value of Load Impedance},
+volume = {170},
+journal = {Procedia Engineering},
+doi = {10.1016/j.proeng.2017.03.064}
 }
