+++ paper/paper.md
@@ -17,56 +17,194 @@ authors:
 affiliations:
  - name: Royal Military Academy, Brussels, Belgium
    index: 1
- - name: Kuleuven, Leuven, Belgium
+ - name: KU Leuven, Leuven, Belgium
    index: 2
 date: August 2024
 bibliography: paper.bib
 ---
 
 # Summary
-The growing intricacy of contemporary engineering systems, typically reduced to stiff differential equations with events, poses a difficulty in digitally simulating them using traditional numerical integration techniques. The Quantized State System (QSS) and the Linearly Implicit Quantized State System (LIQSS) are different methods for tackling such problems. It is an approach that builds the solution by updating the system variables independently as opposed to classic integration methods that update all the system variables every step. [@QSS]. 
+
+Contemporary engineering systems, such as electrical circuits, mechanical systems with shocks, and chemical reactions with rapid kinetics, are often characterized by dynamics that can be modeled using stiff differential equations with events. Stiffness typically arises in these systems due to the presence of both rapidly changing and slowly changing components. This stiffness requires extremely small time steps to maintain stability when using traditional numerical integration techniques. Recently, quantization-based techniques have emerged as an effective alternative for handling such complex models. Methods like the Quantized State System (QSS) and the Linearly Implicit Quantized State System (LIQSS) offer promising results, particularly for large sparse stiff models. Unlike classic numerical integration methods, which update all system variables at each time step, the quantized approach updates individual system variables independently. Specifically, in quantized methods, each variable is updated only when its value changes by a predefined quantization level resulting in only updating the rapidly changing components. Moreover, these methods are advantageous when dealing with discontinuous events, where traditional integrators may struggle with accuracy. An event is a discontinuity where the state of the system abruptly changes at a specific point. Classic methods either undergo expensive iterations to pinpoint the exact discontinuity instance or resort to interpolating its location, resulting in unreliable outcomes. Therefore, this QSS strategy can significantly reduce computational effort and improve efficiency in large sparse stiff models with frequent discontinuities [@QSS].
 
 # Statement of need
-Traditional solvers are challenged by frequent discontinuities where the state of the system abruptly changes
-at specific points or intervals. They struggle to accurately capture the dynamics around discontinuities. They either undergo expensive iterations to pinpoint exact discontinuity instances or resort to interpolating their locations, resulting in unreliable outcomes. 
-Written in the easy-to-learn Julia language [Julia programming language](https://julialang.org) [@julia], 
-and taking advantage of its features such as multiple dispatch and metaprogramming, the QuantizedSystemSolver.jl is a solver that aims to efficiently solve a set of Ordinary differential Equations with a set of events via implementing the QSS and LIQSS methods. It is the first such tool to be published in the Julia ecosystem.
+
[…]
