+++ build/Common/QSSNL_AbstractTypes.jl
@@ -1,11 +1,0 @@
-
-abstract type NLODEProblem{PRTYPE,T,Z,Y,CS} end
-abstract type ALGORITHM{N,O} end
-abstract type Sol{T,O} end
-abstract type SpecialQSS_data{O1,Lightness} end
-abstract type SpecialLiqssQSS_data end
-
-
-
-
-

[…]
