+++ paper/paper.bib
@@ -55,6 +55,7 @@ doi = {10.1177/0037549714536255}
    pages = {15-24},
  year = {2017},
 journal = {The Journal of Open Research Software},
+doi = {10.5334/jors.151},
 }
  
 @Inproceedings{elbellili, 
@@ -71,7 +72,7 @@ journal = {The Journal of Open Research Software},
   publisher = {Springer},
   year = {2006},
   language = {en},
-  address = {simulation.New York},
+  address = {simulation. New York},
 }
 
 @article{buck,
