+++ paper/paper.md
@@ -36,17 +36,18 @@ Written in the easy-to-learn [Julia programming language](https://julialang.org)
 The main idea behind QSS methods is to divide the system state space into quantized regions and represent the system state in terms of these quantized values. The QSS methods update the system status only when certain thresholds or conditions are met, instead of continuously simulating the system's behavior. For instance, state variables with large gradients will get to their thresholds more frequently than states with small gradients. If a state or input gets to its next threshold, a discrete event is triggered, and information is passed on to other integrators that depend on it [@fe2006a].
 The general form of a problem composed of a set of ODEs and a set of events that QSS is able to solve is described in the following: 
 
-```math
-\dot X=f(X,P,t) 
 
- if \; zc_v(x_i...,p_d...,t) \; i \in [1,n]  \;  ; \; d  \in [1,m] 
 
- \qquad x_i=H_v(x_i...,p_d...,t) 
+$\dot X=f(X,P,t) $
 
- \qquad p_d=L_v(x_i...,p_d...,t)  
+$ if \; zc_v(x_i...,p_d...,t) \; i \in [1,n]  \;  ; \; d  \in [1,m] $
 
- \qquad \qquad...
-```
+$ \qquad x_i=H_v(x_i...,p_d...,t) $
+
+$ \qquad p_d=L_v(x_i...,p_d...,t)  $
+
+$ \qquad \qquad...$
+
 where $X = [x_1,x_2...,x_n]^T$ is the state vector, $f:\mathbb{R}^n \rightarrow \mathbb{R}^n$ is the derivative function, and $t$ is the independent variable. $D = [d_1,d_2...,d_m]^T$ is the vector of the system discrete variables. $n$ and $m$ are the number of state variables and discrete variables of the system respectively. $v$ is the number of events and $zc$ is an event condition, $H$ and $L$ are functions used in the effects of the event $zc$.
 
 In classic methods, the difference between $t_k$ (the current time) and $t_{k+1}$ (the next time) is called the step size. 
@@ -95,15 +96,12 @@ The Buck is a converter that decreases voltage and increases current with a grea
 ![The buck converter](buck.png)
 
 The diode $D$ and the switch $S$ can be modeled as two variables resistors $RD$ and $RS$. When the diode and the switch are ON, $RD$ and $RS$ are set to $10^{-5}$, and when they are OFF, they are set to $10^{5}$. The diode is ON when its current is positive. The switch is controlled by the voltage $SC$, and it is ON for $0.5$ x $10^{-4}$ seconds. A mesh and a nodal analysis give the relationship between the different components in the circuit as shown in Eq.(1):
+$ i_d = \frac{RS.i_l-V1}{RS+RD} $
 
-```math
-  i_d = \frac{RS.i_l-V1}{RS+RD} , \\ 
-       .\\
-  \frac{du_c}{dt} = \frac{i_l-\frac{u_c}{R}}{C}  , \\ 
-       .\\
-  \frac{di_l}{dt} = \frac{-uc-i_d.RD}{L}
+$ \frac{du_c}{dt} = \frac{i_l-\frac{u_c}{R}}{C}$
 
-```
+ $ \frac{di_l}{dt} = \frac{-uc-i_d.RD}{L}$
+
 
 The buck problem can be solved by the QuantizedSystemSolver.jl package using the following code:
 
@@ -147,24 +145,21 @@ plot(sol)
 
 
[…]
