+++ paper/paper.bib
@@ -29,7 +29,7 @@
   url = {https://github.com/symengine/SymEngine.jl},
   year = {2015},
 }
-@misc{QSS,
+@misc{improveliqss,
   author       = {Pietro, F. and Migoni, G. and Kofman, E.},
   title        = {Improving linearly implicit quantized state system methods},
   journal = {Simulation: Transactions of the Society for Modeling and Simulation International},
