+++ Project.toml
@@ -10,12 +10,9 @@ Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ExprTools = "e2ba6199-217a-4e67-a87a-7c52f15ade04"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-PolynomialRoots = "3a141323-8675-5d76-9d11-e1df1406c778"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 SymEngine = "123dc426-2d89-5057-bbad-38513e3affd8"
-TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 
 [compat]
 CodeTracking = "1.3.6"
@@ -23,7 +20,6 @@ Documenter = "1.4.1"
 ExprTools = "0.1.10"
 MacroTools = "0.5.13"
 Plots = "1.40"
-PolynomialRoots = "1"
 RuntimeGeneratedFunctions = "0.5.12"
 StaticArrays = "1"
 SymEngine = "0.11"
