### 2024-12-08 4341fc7
more lowercase_uppercase
