+++ paper/paper.md
@@ -30,7 +30,7 @@ Contemporary engineering systems, such as electrical circuits, mechanical system
 # Statement of need
 
 Traditional solvers are challenged by large sparse stiff models and systems with frequent discontinuities. As an example, the Advection Diffusion Reaction (ADR) problem is a large sparse stiff system that classic methods can not efficiently solve. In addtion, the buck converter is a stiff system with frequent discontinuities that classic solvers from the DifferentialEquations.jl [@Rackauckas2017] are currently unable to handle properly.  
-Written in the easy-to-learn [Julia programming language](https://julialang.org) [@julia], inspired by the [`qss-solver`] ( https://github.com/CIFASIS/qss-solver)[@qssC] written in C, and taking advantage of the Julia features such as multiple dispatch and metaprogramming, the QuantizedSystemSolver.jl package shares the same interface as DifferentialEquations.jl package and aims to efficiently solve a large set of stiff Ordinary Differential Equations (ODEs) with a set of events via implementing the QSS and LIQSS methods. It is the first such tool to be published in the Julia ecosystem. This package is able to easily solve the ADR and the buck converter problems as shown in the examples section.
+Written in the easy-to-learn [Julia programming language](https://julialang.org) [@julia], inspired by the [`qss-solver`]( https://github.com/CIFASIS/qss-solver)[@qssC] written in C, and taking advantage of the Julia features such as multiple dispatch and metaprogramming, the QuantizedSystemSolver.jl package shares the same interface as DifferentialEquations.jl package and aims to efficiently solve a large set of stiff Ordinary Differential Equations (ODEs) with a set of events via implementing the QSS and LIQSS methods. It is the first such tool to be published in the Julia ecosystem. This package is able to easily solve the ADR and the buck converter problems as shown in the examples section.
 
 # Quantization-based techniques
 The main idea behind QSS methods is to divide the system state space into quantized regions and represent the system state in terms of these quantized values. The QSS methods update the system status only when certain thresholds or conditions are met, instead of continuously simulating the system's behavior. For instance, state variables with large gradients will get to their thresholds more frequently than states with small gradients. If a state or input gets to its next threshold, a discrete event is triggered, and information is passed on to other integrators that depend on it [@fe2006a].
@@ -73,11 +73,11 @@ odeprob=ODEProblem(func,u,tspan,p)
 ```
 
 The output of the previous `ODEProblem` function, which is a QSS problem, is passed to a solver function with other configuration arguments
-such as the algorithm type and the tolerance. The solve function dispatches on the given algorithm and start the numerical integration.
+such as the algorithm type and the tolerance. The solve function dispatches on the given algorithm and starts the numerical integration.
 ```julia
 sol = solve(odeprob,algorithm,abstol = ...,reltol = ...)    
 ```
-A the end, a solution object is produced that can be queried, plotted, and error-analyzed.
+At the end, a solution object is produced that can be queried, plotted, and error-analyzed.
 
 ```julia
 sol(0.0005,idxs = 2) # get the value of variable 2 at time 0.0005
@@ -87,7 +87,7 @@ plot(sol)          # plot the solution
 ![The package structure](diagram.png)
 
 In addition, the package contains several shared helper functions used during the integration process by the algorithm such as the scheduler that organizes which variable of the system to update at any specific time of the simulation. 
[…]
