### 2024-12-08 635e3c7
not updating folder names lower upper !!!
