+++ docs/src/guide/userGuide.md
@@ -32,7 +32,7 @@ A the end, a solution object is produced that can be queried, plotted, and error
 For simple examples, head over to the examples section.
 
 ## Solving a Nonlinear ODE Problem with events in Julia
-In this section, we will go through the process of setting up, solving, querying, and plotting a system of ordinary differential equation (ODEs) with events. We will use a buck converter circuit model as an example.
+In this section, we will go through the process of setting up, solving, querying, and plotting a non linear system of ordinary differential equation (ODEs) with events. We will use a buck converter circuit model as an example.
 
 #  <span style="color:red">Example: Buck circuit</span>
 [The Buck](https://en.wikipedia.org/wiki/Buck_converter) is a converter that decreases voltage and increases current with a greater power efficiency than linear regulators. Its circuit is shown in the following figure:
