### 2024-12-08 b7e7548
Update userGuide.md
