+++ README.md
@@ -43,10 +43,10 @@ The general form of a problem composed of a set of ODEs and a set of events that
 ```
 where $X = [x_1,x_2...,x_n]^T$ is the state vector, $f:\mathbb{R}^n \rightarrow \mathbb{R}^n$ is the derivative function, and $t$ is the independent variable. $D = [d_1,d_2...,d_m]^T$ is the vector of the system discrete variables. $n$ and $m$ are the number of state variables and discrete variables of the system respectively. $v$ is the number of events and $zc$ is an event condition, $H$ and $L$ are functions used in the effects of the event $zc$.
 
-For new users, take a look at the [Tutorial](@ref) section. If you see something wrong,
+For new users, take a look at the [Tutorial](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/guide/userTutorial/) section. If you see something wrong,
 please open an [issue](https://github.com/mongibellili/QuantizedSystemSolver.jl/issues)
 
-For developpers, take a look at the [Developer Guide](@ref) section. Then, if you have an idea,
+For developpers, take a look at the [Developer Guide](https://mongibellili.github.io/QuantizedSystemSolver.jl/dev/developer/devIntro/) section. Then, if you have an idea,
 do a [pull request](https://github.com/mongibellili/QuantizedSystemSolver.jl/pulls)!
 
 
