+++ paper/paper.md
@@ -175,9 +175,11 @@ function adr(du,u,p,t)
   #Equations
   du[1] = -a*_dx*(u[1]-0.0)+d*_dx*_dx*(u[2]-2.0*u[1]+0.0)+r*u[1]*u[1]*(1.0-u[1]) 
   for k in 2:999  
-      du[k]=-a*_dx*(u[k]-u[k-1])+d*_dx*_dx*(u[k+1]-2.0*u[k]+u[k-1])+r*u[k]*u[k]*(1.0-u[k]) ;
+      du[k]=-a*_dx*(u[k]-u[k-1])+d*_dx*_dx*(u[k+1]-2.0*u[k]+u[k-1])+
+             r*u[k]*u[k]*(1.0-u[k]) ;
   end 
-  du[1000]=-a*_dx*(u[1000]-u[999])+d*_dx*_dx*(2.0*u[999]-2.0*u[1000])+r*u[1000]*u[1000]*(1.0-u[1000]) 
+  du[1000]=-a*_dx*(u[1000]-u[999])+d*_dx*_dx*(2.0*u[999]-2.0*u[1000])+
+            r*u[1000]*u[1000]*(1.0-u[1000]) 
 end
 tspan = (0.0,10.0)
 u0=zeros(1000)
