+++ docs/src/examples/linearTimeInvariantSystem.md
@@ -14,12 +14,11 @@ p1 = plot(sol);
 ```
 Then use DifferentialEquations instead of QuantizedSystemSolver and use solvers such as ABDF2(), QNDF2(), QBDF2(), or Trapezoid() instead of nmliqss2. The plots are displayed below:
 
-<div>
-<img style="  width: 300px;"  src="../../assets/img/lti10_nmliqss2_.png">
-<img style="  width: 300px;"  src="../../assets/img/lti_ABDF2_.png">
-</div>
 
 
+![buck circuit plot](../assets/img/lti10_nmliqss2_.png)
+
+![buck circuit plot](../assets/img/lti_ABDF2_.png)
 
 This is a great example that shows when we need to use the explicit qss, the implicit liqss, or the modified implicit nmliqss. This is a stiff problem so we need to use the implicit methods, but it also contains larger entries outside the main diagonal of the Jacobian. Therefore, nmliqss should the most appropriate algorithm to use. 
 
