### 2024-12-09 mongibellili/QuantizedSystemSolver.jl#11 @mongibellili
- redundant and unnecessary folders and files are removed
- unnecessary comments inside files are removed
- files and folders are renamed to follow the camelCase style
- all functions are now documented with docstrings
