### 2024-12-09 mongibellili/QuantizedSystemSolver.jl#12 @mongibellili
Problem definition is updated to match the differentialEquations.jl interface (except for events which are put inside the function instead of callback functions):

function func(du,u,p,t)
  #parameters
  #helper expressions
  #differential equations
  #if-statments for events
end

tspan=(initial_time,final_time)
u = [u1_0,u2_0...]
p = [p1_0,p2_0...]
odeprob=ODEProblem(func,u,tspan,p)

sol= solve(odeprob,algorithm,abstol=...,reltol=...)
