### 2024-12-09 mongibellili/QuantizedSystemSolver.jl#13 @mongibellili
The readme file now introduces:
- the subject
-  the installation,
- the problem the package aims to solve
- usefull links for users (tutorial page) and developpers.
