### 2024-12-09 mongibellili/QuantizedSystemSolver.jl#14 @mongibellili
- The index page has a statement of need and purpose of the package, information about licence, installation, related packages. it also guides users to relevant resources.
- Mathematical descriptions are added to different problems
- Naviagtion is improved by using the side panel.
- High level documentation is provided for users in tutorial and API pages. developers guide is separate.
- Dependencies to packages is minimized and put under developer guide after further reading
- repeated links are removed.
- Extension of problem, algorithm, and solution are now explained via examples.
- Taylor0 documentation is reworked to include a purpose, and an example.
- Examples are put into separate subpages
- Added ```jldoctest ...... ``` to examples in docstrings to ensure that code runs correctly.
- build folder is excluded.
