### 2024-12-09 mongibellili/QuantizedSystemSolver.jl#15 @mongibellili
doc strings are added for all functions
