+++ docs/src/examples/linearTimeInvariantSystem.md
@@ -12,13 +12,26 @@ odeprob = ODEProblem(sysb53,u0,tspan)
 sol = solve(odeprob,nmliqss2())
 p1 = plot(sol);
 ```
+
+The analytic solution is given by the following equations:
+```julia
+ x1(t)=105.83453057676363*-8.83362618880318*exp(-10.943696474115944*t)-127.91533865757171*-7.303470585390368*exp(-9.046303525884058*t)-0.32323232323231316
+ x2(t)=105.83453057676363*exp(-10.943696474115944*t)-127.91533865757171*exp(-9.046303525884058*t)+20.08080808080808
+```
 Then use DifferentialEquations instead of QuantizedSystemSolver and use solvers such as ABDF2(), QNDF2(), QBDF2(), or Trapezoid() instead of nmliqss2. The plots are displayed below:
 
 
+**LTI plot using QuantizedSystemSolver.jl**
 
-![buck circuit plot](../assets/img/lti10_nmliqss2_.png)
+![LTI plot](../assets/img/lti10_nmliqss2_.png)
 
-![buck circuit plot](../assets/img/lti_ABDF2_.png)
+**LTI plot using DifferentialEquations.jl**
+
+![LTI plot](../assets/img/lti_ABDF2_.png)
+
+**LTI plot using the analytic solution**
+
+![LTI plot](../assets/img/plot_analysolltiA.png)
 
 This is a great example that shows when we need to use the explicit qss, the implicit liqss, or the modified implicit nmliqss. This is a stiff problem so we need to use the implicit methods, but it also contains larger entries outside the main diagonal of the Jacobian. Therefore, nmliqss should the most appropriate algorithm to use. 
 

[…]
