### 2024-12-10 mongibellili/QuantizedSystemSolver.jl#16 @mongibellili
Two stiff ODEs with events are added (LTI and buck). The correctness for the LTI with evets is shown through an analytic solution.
The buck correctness is shown via the solution from other tools such as the QSS in C and the ltspice software.
