### 2024-12-10 mongibellili/QuantizedSystemSolver.jl#17 @mongibellili
test files are cleaned and commented-out sections are removed.
comparisons of computed results against reference solutions are added for the tyson and the ADR problems.
camelCase is used for file names.
