### 2024-12-10 d327df8
added CITATION.cff, updated README.md with zenodo
