+++ typeC.jl
@@ -1,37 +1,0 @@
-
- function A1() 
- odeprob = @NLodeProblem begin
-                name=(A1,)
-                u = [-1.0, -2.0]
-                du[1] = -20.0*u[1]+-80.0*u[2]+1600.0
-                du[2] =1.24*u[1]+0.01*u[2]+0.2
-            end  
- x1(t)=105.83453057676363*-8.83362618880318*exp(-10.943696474115944*t)+-127.91533865757171*-7.303470585390368*exp(-9.046303525884058*t)+-0.32323232323231316
- x2(t)=105.83453057676363*exp(-10.943696474115944*t)+-127.91533865757171*exp(-9.046303525884058*t)+20.08080808080808
- return (odeprob,x1,x2) 
- end 
-
-
- function B1() 
- odeprob = @NLodeProblem begin
-                name=(B1,)
-                u = [-0.3168087916501198, 20.079871428218116]
-                du[1] = -22.0*u[1]+-80.0*u[2]+1000.0
-                du[2] =1.24*u[1]+0.01*u[2]+0.2
-            end  
- x1(t)=-5.300798816237515e17*-12.649850456341543*exp(-15.675814565863513*t)+9.014692147675046e7*-5.10014954365846*exp(-6.314185434136489*t)+-0.26267932915740405
- x2(t)=-5.300798816237515e17*exp(-15.675814565863513*t)+9.014692147675046e7*exp(-6.314185434136489*t)+12.572236815518284
- return (odeprob,x1,x2) 
- end 
-
- function C1() 
- odeprob = @NLodeProblem begin
-                name=(C1,)
-                u = [-0.26861422830125536, 12.573418277218007]
-                du[1] = -22.0*u[1]+-20.0*u[2]+1000.0
-                du[2] =1.24*u[1]+-10.0*u[2]+-0.5
-            end  
- x1(t)=4.746452513218909e42*-7.537612988819599*exp(-19.346640106136302*t)+8.75190264440514e27*-2.139806366019111*exp(-12.653359893863698*t)+40.890522875817
- x2(t)=4.746452513218909e42*exp(-19.346640106136302*t)+8.75190264440514e27*exp(-12.653359893863698*t)+5.020424836601308
- return (odeprob,x1,x2) 
- end 
