+++ CITATION.cff
@@ -1,0 +1,11 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Bellili"
+  given-names: "Mongi"
+  orcid: "https://orcid.org/0000-0003-1230-5488"
+title: "Quantized System Solver"
+version: 1.0.3
+doi: 10.5281/zenodo.14361143
+date-released: 2024-12-10
+url: "https://doi.org/10.5281/zenodo.14361143"
