+++ CITATION.cff
@@ -1,8 +1,8 @@
 cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 authors:
-- family-names: "Bellili"
-  given-names: "Mongi"
+- family-names: "ElBellili"
+  given-names: "ElMongi"
   orcid: "https://orcid.org/0000-0003-1230-5488"
 title: "Quantized System Solver"
 version: 1.0.3
