### 2024-12-10 f3aa675
add own doi to paper
