+++ paper/paper.bib
@@ -85,3 +85,14 @@ volume = {170},
 journal = {Procedia Engineering},
 doi = {10.1016/j.proeng.2017.03.064}
 }
+
+
+@software{ElBellili2024,
+author = {ElBellili, ElMongi and Huybrechs, Daan  and Lauwens, Ben},
+doi = {10.5281/zenodo.14361143},
+month = dec,
+title = {{Quantized System Solver}},
+url = {https://doi.org/10.5281/zenodo.14361143},
+version = {1.0.3},
+year = {2024}
+}
