### 2024-12-10 mongibellili/QuantizedSystemSolver.jl#19 @mongibellili
First of all, I am grateful for your time spent to provide these comments and suggestions.
**Preamble**
 changed "Kuleuven" to "KU Leuven"
**Paragraph 1**
- added examples of contemporary engineering systems
- Stiffness explained
- more explanation of the QSS methods is added in the summary as well as in a separate section with a mathematical expression and an algorithm.
**Paragraph 2**
- Event is now defined in the summary.
- grammar and formatting errors are removed
- the sentence of the purpose of the package is reworked.
 **Paragraph 3**
- The terms "Problem," "Algorithm," and "Solution" are now not capitalized, but with a different font.
- The figure is referenced within the text using a figure number.
- NLodeProblem name is changed to ODEProblem similar to the DifferentialEquations.jl
- Code examlpes use Julia code.
- Mathematical equations are provided before the julia code for the the two examples.
- Real julia code is used with comments to highlight the meanings
- The use of quote ... end is replaced by a conventional approach more familiar to julia users.
- whitespace around = is added
- whitespace after, inside code is removed.
- whitespace before citiations is added.
- author name is added in some references
**Figures**
- whitespace around = in figure1 is added
- caption is added
 **Conclusion**
- the conlusion is sepecific about the functionalities offered by the package
- lengthy sentence is reworked
**References**
- doi is added to sources that have them.
**General Remarks**
- Statement of Need: It is mentioned what type of problems classic solvers are unable to handle efficiently. the exact exmples (Buck and ADR) are put in the examples section.
- Comparison with Other Packages: the performance against various classic solvers is added in a figure.
- (https://github.com/CIFASIS/qss-solver)  is referenced since it also implements the QSS methods.
- Ongoing Research and Publications: This package is recent and it is being integrating in another tool (in progess). There is one publication that used this packaged and it was referenced under "Quantization-based techniques" section. Another article enabled by this package is in progress.
- the software is archived using Zenodo and a citation with a DOI is added in the paper
