### 2024-12-10 mongibellili/QuantizedSystemSolver.jl#20 @mongibellili
First of all, I am grateful for your time spent to provide these comments and suggestions.
- Direction to documentation is now clearer.
- Installation is explained in a step by step approach.
- Explanation of some limitations of the classic methods is added, and a comparison via two examples is also added.
- mentioning the capabilities of the different versions is removed from the readme.

**Documentation:**
- the landing page contains the statement of need and a detailed installation steps.
- The difference between QSS and classic methods is stated, and a comparison is added in the examples.
- The syntax to setup the problem and solve it is added before the first exmple in the tutorial section.
- The different subject are split into multiple subpages for an easier navigation, and the user API is separate.
- implementation and dependencies are minimized and kept under the developer guide.
