+++ paper/paper.bib
@@ -62,7 +62,7 @@ doi = {10.5334/jors.151},
   author =        {Elbellili, E. and Lauwens, B. and Huybrechs, D.},
   title =        {Investigation of Different Conditions to Detect Cycles in Linearly Implicit Quantized State Systems},
   booktitle =    {Proceedings of the 3rd. International Conference on Computational Modeling, Simulation and Optimization},
-  year =         {2024},
+  year =         {2024a},
   publisher =    {IEEE Computer Society},
 }
 
@@ -94,5 +94,5 @@ month = dec,
 title = {{Quantized System Solver}},
 url = {https://doi.org/10.5281/zenodo.14361143},
 version = {1.0.3},
-year = {2024}
+year = {2024b}
 }
