+++ docs/src/developer/devIntro.md
@@ -2,7 +2,7 @@
 
 ## Contribution
 
-QuantizedSystemSolver is a collaborative project open to contributions. If you want to contribute, please take into account:
+QuantizedSystemSolver is a collaborative project open to contributions. If you want to contribute, please take these steps into account:
 
   - Carefully read and follow this guide. 
   - Before opening a PR with a significant contribution, contact the project administrators, e.g., by writing an email to bellilimongi@gmail.com or by opening an issue describing what you are willing to implement. Wait for feed-back.
@@ -20,7 +20,7 @@ julia> ]
 (@v1.x) pkg> dev("/some/path/")
 ```
 
-Want to help in other ways? We have a number of [issues](https://github.com/mongibellili/QuantizedSystemSolver.jl/issues). You can start contributing to the project by solving some of those issues. 
+Want to help in other ways? You can start contributing to the project by solving some of the [issues](https://github.com/mongibellili/QuantizedSystemSolver.jl/issues). 
 
 
 
@@ -40,7 +40,7 @@ See [the Julia CONTRIBUTING.md](https://github.com/JuliaLang/julia/blob/master/C
 ## Components overview
 While the package is optimized to be fast, extensibility is not compromised. It is divided into 3 entities that can be extended separately: Problem, Algorithm, and Solution. The package uses other packages such as MacroTools.jl for user-code parsing, SymEngine.jl for Jacobian computation, and a modified TaylorSeries.jl that uses caching to obtain free Taylor variables. 
 The other components of the solver include the NLodeProblem function, the solve function, the scheduler, and the quantizer. The following sections provide an in-depth explanation of the key components and their interactions within the solver framework.
-::: center
+
 ![The QSS Solver Structure.](../assets/img/diagram.png)
 
 ## Future work

[…]
