+++ paper/paper.bib
@@ -68,7 +68,7 @@ doi = {10.5334/jors.151},
 
 @book{fe2006a,
   author = {Cellier, F. and Kofman , E.},
-  title = {Continuous system},
+  title = {Continuous System Simulation},
   publisher = {Springer},
   year = {2006},
   language = {en},
