+++ paper/paper.bib
@@ -75,17 +75,17 @@ doi = {10.5334/jors.151},
   address = {simulation. New York},
 }
 
-@article{buck,
-author = {Cristri, A.W.},
-year = {2017},
-month = {12},
-pages = {398-403},
-title = {Analysis and Design of Dynamic Buck Converter with Change in Value of Load Impedance},
-volume = {170},
-journal = {Procedia Engineering},
-doi = {10.1016/j.proeng.2017.03.064}
-}
 
+@article{buck,
+author = {Migoni, Gustavo and Kofman, Ernesto and Bergero, Federico and Fernandez, Joaquin},
+year = {2015},
+month = {04},
+pages = {320-336},
+title = {Quantization-based simulation of switched mode power supplies},
+volume = {91},
+journal = {SIMULATION},
+doi = {10.1177/0037549715575197}
+}
 
 @software{ElBellili2024,
 author = {ElBellili, ElMongi and Huybrechs, Daan  and Lauwens, Ben},
