+++ src/integrators/qssDiscreteIntegrator.jl
@@ -1,12 +1,12 @@
 """
-    integrate(Al::QSSAlgorithm{:qss,O}, CommonqssData::CommonQSS_Data{Z}, odep::NLODEProblem{PRTYPE,T,D,Z,CS}, f::Function, jac::Function, SD::Function) where {PRTYPE,O,T,D,Z,CS}
+    integrate(Al::QSSAlgorithm{:qss,O}, CommonqssData::CommonQSS_Data{Z}, odep::NLODEProblem{F,PRTYPE,T,D,Z,CS}, f::Function, jac::Function, SD::Function) where {F,PRTYPE,O,T,D,Z,CS}
 
 Integrates a nonlinear ordinary differential equation (ODE) problem with `events` using a Quantized State System (QSS) algorithm.
 
 # Arguments
 - `Al::QSSAlgorithm{:qss,O}`: The QSS algorithm to be used for integration.
 - `CommonqssData::CommonQSS_Data{Z}`: Common data structure for QSS algorithms.
-- `odep::NLODEProblem{PRTYPE,T,D,Z,CS}`: The nonlinear ODE problem to be solved.
+- `odep::NLODEProblem{F,PRTYPE,T,D,Z,CS}`: The nonlinear ODE problem to be solved.
 - `f::Function`: The function defining the ODE system.
 - `jac::Function`: The Jacobian dependency function of the ODE system.
 - `SD::Function`: The state derivative dependency function.
@@ -23,7 +23,7 @@ Integrates a nonlinear ordinary differential equation (ODE) problem with `events
 # Returns
 - A solution
 """
-function integrate(Al::QSSAlgorithm{:qss,O}, CommonqssData::CommonQSS_Data{Z}, odep::NLODEProblem{PRTYPE,T,D,Z,CS}, f::Function, jac::Function, SD::Function) where {PRTYPE,O,T,D,Z,CS}
+function integrate(Al::QSSAlgorithm{:qss,O}, CommonqssData::CommonQSS_Data{Z}, odep::NLODEProblem{F,PRTYPE,T,D,Z,CS}, f::Function, jac::Function, SD::Function) where {F,PRTYPE,O,T,D,Z,CS}
   if VERBOSE println("integration...") end
   ft = CommonqssData.finalTime
   initTime = CommonqssData.initialTime
@@ -44,6 +44,7 @@ function integrate(Al::QSSAlgorithm{:qss,O}, CommonqssData::CommonQSS_Data{Z}, o
   savedTimes = CommonqssData.savedTimes
   taylorOpsCache = CommonqssData.taylorOpsCache
   d = CommonqssData.d
+  clF=odep.closureFuncs[1]
   zc_SimpleJac = odep.ZCjac
   HZ = odep.HZ
   HD = odep.HD
@@ -59,7 +60,7 @@ function integrate(Al::QSSAlgorithm{:qss,O}, CommonqssData::CommonQSS_Data{Z}, o
    end # q computed from x and it is going to be used in the next x
    for i = 1:T
      clearCache(taylorOpsCache, Val(CS), Val(O))
-      f(i, -1, -1, q, d, t, taylorOpsCache)
+      f(i, -1, -1, q, d, t, taylorOpsCache,clF)
        if k==1
          x[i].coeffs[k+1] = (taylorOpsCache[1].coeffs[1])  # /fact cuz i will store der/fac like the convention...to extract the derivatives (at endof sim) multiply by fac  derderx=coef[3]*fac(2)
        elseif k==2
@@ -84,7 +85,7 @@ function integrate(Al::QSSAlgorithm{:qss,O}, CommonqssData::CommonQSS_Data{Z}, o
       computeNextTime(Val(O), i, initTime, nextInputTime, x, quantum)
[…]
