+++ src/problem/qssProblemDefinition.jl
@@ -1,5 +1,5 @@
 """
-    NLODEContProblem{PRTYPE,T,D,Z,CS}
+    NLODEContProblem{F,PRTYPE,T,D,Z,CS}
 A struct that holds the continuous problem. It has the following fields:  
   - `prname`: The name of the problem  
   - `prtype`: The type of the problem  
@@ -13,8 +13,9 @@ A struct that holds the continuous problem. It has the following fields:
   - `jac`: The Jacobian dependency  
   - `SD`: The state derivative dependency  
   - `exactJac`: The exact Jacobian function  
+  - `closureFuncs::Vector{F}` # function that holds closure function inside system defined by user
 """
-struct NLODEContProblem{PRTYPE,T,D,Z,CS}<: NLODEProblem{PRTYPE,T,D,Z,CS} 
+struct NLODEContProblem{F,PRTYPE,T,D,Z,CS}<: NLODEProblem{F,PRTYPE,T,D,Z,CS} 
   prname::Symbol # problem name used to distinguish printed results
   prtype::Val{PRTYPE} # problem type: not used but created in case in the future we want to handle problems differently
   a::Val{T} #problem size based on number of vars: T is used not a: 'a' is a mute var
@@ -27,9 +28,10 @@ struct NLODEContProblem{PRTYPE,T,D,Z,CS}<: NLODEProblem{PRTYPE,T,D,Z,CS}
   jac::Vector{Vector{Int}}#Jacobian dependency..I have a der and I want to know which vars affect it...opposite of SD...is a vect for direct method (later @resumable..closure..for saved method)
   SD::Vector{Vector{Int}}#  I have a var and I want the der that are affected by it
   exactJac::Function  # used only in the implicit intgration
+  closureFuncs::Vector{F} # function that holds closure function inside system defined by user
 end
 """
-    NLODEContProblemSpan{PRTYPE,T,D,Z,CS}
+    NLODEContProblemSpan{F,PRTYPE,T,D,Z,CS}
 A struct that holds the continuous problem with tspan. It has the following fields:  
   - `prname`: The name of the problem  
   - `prtype`: The type of the problem  
@@ -44,8 +46,9 @@ A struct that holds the continuous problem with tspan. It has the following fiel
   - `SD`: The state derivative dependency  
   - `exactJac`: The exact Jacobian function  
   - `tspan::Tuple{Float64, Float64}`:  This field variable did not exist in the original NLODEContProblem as this simulation time should part of the problem. However, to match the differentialEqation.jl interface, the tspan is added to the definition of the problem.
+  - `closureFuncs::Vector{F}` # function that holds closure function inside system defined by user
 """
-struct NLODEContProblemSpan{PRTYPE,T,D,Z,CS}<: NLODEProblem{PRTYPE,T,D,Z,CS} 
+struct NLODEContProblemSpan{F,PRTYPE,T,D,Z,CS}<: NLODEProblem{F,PRTYPE,T,D,Z,CS} 
   prname::Symbol # problem name used to distinguish printed results
   prtype::Val{PRTYPE} # problem type: not used but created in case in the future we want to handle problems differently
   a::Val{T} #problem size based on number of vars: T is used not a: 'a' is a mute var
@@ -59,11 +62,27 @@ struct NLODEContProblemSpan{PRTYPE,T,D,Z,CS}<: NLODEProblem{PRTYPE,T,D,Z,CS}
   SD::Vector{Vector{Int}}#  I have a var and I want the der that are affected by it
   exactJac::Function  # used only in the implicit intgration
[…]
