### 2025-01-17 mongibellili/QuantizedSystemSolver.jl#19 @lamBOOO
Thank you for addressing my points. Regarding the paper, I have some follow up points in order for the paper to meet the standards of JOSS.

- The word count is over 1800. According to [the guidelines here](https://joss.readthedocs.io/en/latest/paper.html), papers should be between 250-1000. So a reduction of the paper length is required. I try to give some recommendations for shortening below.
- In the title "QuantizedSystemSolver: A discontinuous ODE system solver for Julia", the wording "for Julia" sounds a bit strange. Are you sure that this is correct? Shouldn't it be "in Julia"?
- For the code archive, the "Archive" link to the top-left of the paper header should be used instead of a citation in the text (line 32).
- The beginning of "Quantization-based techniques" section has some repetition, this might be shortened or removed.
- The description of the mLIQSS1 algorithm needs a lot of space. I would like to have it shortened. Maybe converting it to a high quality figure would save the space when the figure spans the whole horizontal space. I'm not even sure if this level of detail is necessary in the scope of JOSS.
- The "if" in line 48 should not be italic since it's not a variable.
- line 52 has f:R^n -> R^n but f takes X in R^n, P in R^m and t in R+, so it should rather be f:(R^n)x(R^m)x(R+) -> R^n, right?
- You used separate equations (line 47-51) with `$...$`. It would be better (imo), the used an aligned env. I.e., replacing
```
$\dot X = f(X,P,t)$

$if \; zc_v(x_i...,p_d...,t) \; i \in [1,n]  \;  ; \; d  \in [1,m]$

$\qquad x_i = H_v(x_i...,p_d...,t)$

$\qquad p_d = L_v(x_i...,p_d...,t)$

$\qquad \qquad...$
```
with
```
$$\begin{align}
\dot X = f(X,P,t)
\\
if \; zc_v(x_i...,p_d...,t) \; i \in [1,n]  \;  ; \; d  \in [1,m]
\\
\qquad x_i = H_v(x_i...,p_d...,t)
\\
\qquad p_d = L_v(x_i...,p_d...,t)
\\
\qquad \qquad...
\end{align}$$
```
- I find the notation is line 47-51 a little bit complicated, wouldn't something like that be simpler having the same meaning?
$$\begin{align}
\dot X = f(X,P,t)
\\
\textnormal{if} \; \textnormal{zc}_v(X,P,t): \; \textnormal{Set }
x_i = H_v(X,P,t)
\textnormal{ and }
p_d = L_v(X,P,t)
\end{align}$$

- What is the meaning of the `...` in line 51?
- For what stands the `v` in the event $zc_v$?
- The mLIQSS1 algorithm, additionally, is quite detailed. Maybe save space by using pseudo code to just explain the most important steps.
- Please use consistent code styling. E.g., in line 92-93, there is not space behind the `#` starting a comment while in the code block after line 98, there is a space. I would suggest adding a space after each `#` comments.
-  the code after line 92 can be shrinked and improved by including an actual example, maybe something like that:
```julia
# parameters, helpers, differential eqs., if-statements for events; e.g.:
du[1] = p[1] * u[1]
if (t > 1) p[1] = 42.0 end
```
- I'm unsure if the Fig. 1 is even beneficial for the understanding, since it's basically just the API, which could also be accessed through the documentation. With the minimal working example (MWE) from before, the structure of the package should already be clear in my opinion.
- Although helpful, I find the Examples section now a bit long. Is it possible to maybe only focus one example and try to "compress" its code s.t. the paper doesn't get too long?
	- Maybe `du[1]=..` and `du[2]=..` on the same line, if-statements in one line.
- Please also use consistent coding style in the block after line 122 (spacing after ";", spacing after "#", whitespace around "=" and ">", etc.)
- The quality of Fig. 3 is not optimal. It is quite pixely. Please try to improve the figure and the layout (by combining it, maybe, with another figure side-by-side)
- The code for the ADR is not so easy to read with all the indices, maybe the buck converter example is more helpful in showing what the software is, in principle, about without confusing.
- I like the CPU time comparison, is it a fair comparison with the 2nd-order stiff solvers (in terms of parameters to set)?
- "Rosenbrock23" is missing the "()" in comparison to the other methods.
- The contents of Fig. 5 would also be suited for a table, also saving space if the solvers are put on the "x-axis" of the table.
