+++ paper/paper.md
@@ -1,7 +1,7 @@
 ---
-title: 'QuantizedSystemSolver: A discontinuous ODE system solver for Julia.'
+title: 'QuantizedSystemSolver: A discontinuous ODE system solver in Julia.'
 tags:
-  - Quantised State system
+  - Quantised State System
   - Discontinuities
   - Stiff Ordinary Differential Equations
 authors:
@@ -29,127 +29,64 @@ Contemporary engineering systems, such as electrical circuits, mechanical system
 
 # Statement of need
 
-Traditional solvers are challenged by large sparse stiff models and systems with frequent discontinuities. As an example, the Advection Diffusion Reaction (ADR) problem is a large sparse stiff system that classic methods cannot efficiently solve. In addition, the buck converter is a stiff system with frequent discontinuities that classic solvers from the DifferentialEquations.jl [@Rackauckas2017] are currently unable to handle properly.  
-Written in the easy-to-learn Julia programming language [@julia] and inspired by the ``qss-solver`` written in C [@qssC], the QuantizedSystemSolver.jl package [@ElBellili2024] takes advantage of Julia features such as multiple dispatch and metaprogramming. The package shares the same interface as the DifferentialEquations.jl package and aims to efficiently solve a large set of stiff Ordinary Differential Equations (ODEs) with events by implementing the QSS and LIQSS methods. It is the first such tool to be published in the Julia ecosystem. This package can easily solve the ADR and buck converter problems, as shown in the Examples Section.
+Traditional solvers are challenged by large sparse stiff models and systems with frequent discontinuities. The buck converter is a stiff system with frequent discontinuities that classic solvers from the DifferentialEquations.jl [@Rackauckas2017] are currently unable to handle properly.  
+Written in the easy-to-learn Julia programming language [@julia] and inspired by the ``qss-solver`` written in C [@qssC], the QuantizedSystemSolver.jl package takes advantage of Julia features such as multiple dispatch and metaprogramming. The package shares the same interface as the DifferentialEquations.jl package and aims to efficiently solve a large set of stiff Ordinary Differential Equations (ODEs) with events by implementing the QSS and LIQSS methods. It is the first such tool to be published in the Julia ecosystem. 
 
 # Quantization-based techniques
-The main idea behind QSS methods is to divide the system state space into quantized regions and represent the system state in terms of these quantized values. The QSS methods update the system state only when certain thresholds or conditions are met, instead of continuously simulating the system's behavior. For instance, state variables with large gradients will reach their thresholds more frequently than states with small gradients. If a state or input gets to its next threshold, a discrete event is triggered, and information is passed on to other integrators that depend on it [@fe2006a].
+
 The general form of a problem composed of a set of ODEs and a set of events that QSS is able to solve is described in the following equations: 
-
-
 
 $\dot X = f(X,P,t)$
 
-$if \; zc_v(x_i...,p_d...,t) \; i \in [1,n]  \;  ; \; d  \in [1,m]$
+if $zc(X,P,t)$ ; Set $x_i=H(X,P,t)$ and $p_j=L(X,P,t)$
 
-$\qquad x_i = H_v(x_i...,p_d...,t)$
[…]
