+++ src/problem/qssProblemDiscrete.jl
@@ -153,7 +153,6 @@ function NLodeProblemFunc(odeExprs::Expr,::Val{T},::Val{D},::Val{Z},initCond::Ve
     #closurefunc=()->nothing
     closurefunc=@RuntimeGeneratedFunction(functionEx.args[2]) 
     diffEqfunction=createDiscEqFun(otherCode,equs,zcequs,eventequs,fname,closurefunc)# diff equations before this are stored in a dict:: now we have a giant function that holds all diff equations
-    
     else
         closurefunc=0
         diffEqfunction=createDiscEqFun(otherCode,equs,zcequs,eventequs,fname,closurefunc)# diff equations before this are stored in a dict:: now we have a giant function that holds all diff equations
