+++ paper/paper.md
@@ -36,15 +36,15 @@ Written in the easy-to-learn Julia programming language [@julia] and inspired by
 
 The general form of a problem composed of a set of ODEs and a set of events that QSS is able to solve is described in the following equations: 
 
-$\dot X = f(X,P,t)$
-
-if $zc(X,P,t)$ ; Set $x_i=H(X,P,t)$ and $p_j=L(X,P,t)$
+$\dot X = f(X,P,t)$; if $zc(X,P,t)$ ; Set $x_i=H(X,P,t)$ and $p_j=L(X,P,t)$
 
 where $X = [x_1,x_2...,x_n]^T$ is the state vector, $f:\mathbb{R}^n.\; {R}^m. \;{R}^+ \rightarrow \mathbb{R}^n$ is the derivative function, and $t$ is the independent variable. $P = [p_1,p_2...,p_m]^T$ is the vector of the system discrete variables. $n$ and $m$ are the number of state variables and discrete variables of the system respectively. $zc$ is an event condition, $H$ and $L$ are functions used in the effects of the event $zc$.
 
 In QSS, besides the step size, the difference between $x_i(t_k)$ (the current value) and $x_i(t_{k+1})$ (the next value) is called the quantum $\Delta_i$. Depending on the type of the QSS method (explicit or implicit), a new variable $q_i$ is set to equal $x_i(t_k)$  or $x_i(t_{k+1})$ respectively. $q_i$ is called the quantized state of $x_i$, and it is used in updating the derivative function [@elbellili].  A general description of a QSS algorithm is given as follows:
 ![](alg.png)
+
 # Package description
+
 While the package is optimized to be fast, extensibility is not compromised. It is divided into three entities that can be extended separately: The ``problem``, the ``algorithm``, and the ``solution``. The rest of the code creates these entities and glue them together. The API was designed to match the differentialEquations.jl interface while providing an easier way to handle events. The problem is defined inside a function, in which the user may introduce any parameters, variables, equations, and events:
 ```julia
 function func(du,u,p,t) 
@@ -76,6 +76,7 @@ plot(sol)          # plot the solution
 The solver uses other packages such as  [`MacroTools.jl`]( https://github.com/FluxML/MacroTools.jl) [@MacroTools] for user-code parsing, [`SymEngine.jl`]( https://github.com/symengine/SymEngine.jl) [@SymEngine] for Jacobian computation and dependency extraction. It also uses and a modified [`TaylorSeries.jl`](https://github.com/JuliaDiff/TaylorSeries.jl/) [@TaylorSeries] that implements caching to obtain free Taylor variable operations, since the current version of TaylorSeries creates a heap allocated object for every operation. The approximation through Taylor variables transforms any complicated equations to polynomials, making root finding cheaper--a process that QSS methods rely on heavily. 
 
 # The buck converter example
+
 The Buck is a converter that decreases voltage and increases current with a greater power efficiency than linear regulators [@buck]. Its circuit is shown in Fig.1(a).
 
 ![The buck converter](buck.png)
@@ -92,7 +93,7 @@ The buck problem contains frequent discontinuities and can be solved by the Quan
 using QuantizedSystemSolver
 function buck(du,u,p,t)
   #Constant parameters
-  C = 1e-4; L = 1e-4; R = 10.0; V1 = 24.0; T = 1e-4; DC = 0.5; ROn = 1e-5; ROff = 1e5
+  C = 1e-4; L = 1e-4; R = 10.0; V1 = 24.0; T = 1e-4; ROn = 1e-5; ROff = 1e5
[…]
