+++ .github/workflows/CI.yml
@@ -14,7 +14,6 @@ jobs:
       matrix:
         version:
           - '1.6'
-          - '1'
         os:
           - ubuntu-latest
         arch:
