### 2025-01-21 mongibellili/QuantizedSystemSolver.jl#19 @lamBOOO
Hi @mongibellili,

the paper looks good to me now.

I have some last editorial remarks (typos, formatting, ...) collected in: https://github.com/mongibellili/QuantizedSystemSolver.jl/pull/23
