### 2025-01-21 mongibellili/QuantizedSystemSolver.jl#19 @lamBOOO
Some last things: https://github.com/mongibellili/QuantizedSystemSolver.jl/pull/24
