### 2025-01-22 #7434 @diehlpk
Hi @mongibellili please check the invalid DOI reported above and might fix it.
