+++ paper/paper.bib
@@ -34,7 +34,7 @@
   title        = {Improving linearly implicit quantized state system methods},
   journal = {Simulation: Transactions of the Society for Modeling and Simulation International},
   year = {2019},
-  doi     = {10.1177/0037549718766},
+  doi     = {10.1177/0037549718766689},
 }
 
 @article{qssC,
