### 2025-01-22 #7434 @mongibellili
@diehlpk

- Authors are checked for orcid and afiliations
- v1.0.2
- https://doi.org/10.5281/zenodo.14361143
- The title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- MIT license is used for both.

thank you for your time!
