### 2025-01-23 #7434 @diehlpk
@mongibellili it seems that Version v1.0.3 was uploaded to Zenodo, so I used that version as well. Please double check.
