### 2025-01-23 7dc7e44
minor changes in paper.md
