+++ paper/paper.bib
@@ -50,7 +50,7 @@ doi = {10.1177/0037549714536255}
 
 @article{Rackauckas2017,
  author = {Rackauckas, C. and Nie, Q.},
- title = {DifferentialEquations.jl A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+ title = {DifferentialEquations.jl A Performant and Feature-Rich Ecosystem for Solving Differential Equations in {J}ulia},
   volume = {5},
    pages = {15-24},
  year = {2017},
