### 2025-01-23 #7434 @danielskatz
@mongibellili - as track editor, I've proofread the paper, and suggested a few minor changes in https://github.com/mongibellili/QuantizedSystemSolver.jl/pull/25 - please merge this or let me know what you disagree with, then we can proceed to publication
