### 2025-01-23 #7434 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
