### 2025-01-23 #7434 @danielskatz
Congratulations to @mongibellili (elmongi elbellili) and co-authors on your publication!!

JOSS depends on volunteers, and we couldn't do this without your work.
